# chenoj-frontend/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "接口文档",
    "contact": {}
  },
  "host": "localhost:8121",
  "basePath": "/api",
...
  "paths": {
...
    "/api/question/add": {
      "post": {
        "tags": [
          "question-controller"
        ],
        "summary": "addQuestion",
        "operationId": "addQuestionUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "questionAddRequest",
            "description": "questionAddRequest",
            "required": true,
            "schema": {
              "originalRef": "QuestionAddRequest",
              "$ref": "#/definitions/QuestionAddRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«long»",
              "$ref": "#/definitions/BaseResponse«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«long»",
              "$ref": "#/definitions/BaseResponse«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/question/delete": {
      "post": {
        "tags": [
          "question-controller"
        ],
        "summary": "deleteQuestion",
        "operationId": "deleteQuestionUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "deleteRequest",
            "description": "deleteRequest",
            "required": true,
            "schema": {
              "originalRef": "DeleteRequest",
              "$ref": "#/definitions/DeleteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/question/edit": {
      "post": {
        "tags": [
          "question-controller"
        ],
        "summary": "editQuestion",
        "operationId": "editQuestionUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "questionEditRequest",
            "description": "questionEditRequest",
            "required": true,
            "schema": {
              "originalRef": "QuestionEditRequest",
              "$ref": "#/definitions/QuestionEditRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/question/get": {
      "get": {
        "tags": [
          "question-controller"
        ],
        "summary": "getQuestionById",
        "operationId": "getQuestionByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Question»",
              "$ref": "#/definitions/BaseResponse«Question»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Question»",
              "$ref": "#/definitions/BaseResponse«Question»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/question/get/vo": {
      "get": {
        "tags": [
          "question-controller"
        ],
        "summary": "getQuestionVOById",
        "operationId": "getQuestionVOByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«QuestionVO»",
              "$ref": "#/definitions/BaseResponse«QuestionVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«QuestionVO»",
              "$ref": "#/definitions/BaseResponse«QuestionVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/question/list/page": {
      "post": {
        "tags": [
          "question-controller"
        ],
        "summary": "listQuestionByPage",
        "operationId": "listQuestionByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "questionQueryRequest",
            "description": "questionQueryRequest",
            "required": true,
            "schema": {
              "originalRef": "QuestionQueryRequest",
              "$ref": "#/definitions/QuestionQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«Question»»",
              "$ref": "#/definitions/BaseResponse«Page«Question»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«Question»»",
              "$ref": "#/definitions/BaseResponse«Page«Question»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/question/list/page/vo": {
      "post": {
        "tags": [
          "question-controller"
        ],
        "summary": "listQuestionVOByPage",
        "operationId": "listQuestionVOByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "questionQueryRequest",
            "description": "questionQueryRequest",
            "required": true,
            "schema": {
              "originalRef": "QuestionQueryRequest",
              "$ref": "#/definitions/QuestionQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«QuestionVO»»",
              "$ref": "#/definitions/BaseResponse«Page«QuestionVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«QuestionVO»»",
              "$ref": "#/definitions/BaseResponse«Page«QuestionVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/question/my/list/page/vo": {
      "post": {
        "tags": [
          "question-controller"
        ],
        "summary": "listMyQuestionVOByPage",
        "operationId": "listMyQuestionVOByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "questionQueryRequest",
            "description": "questionQueryRequest",
            "required": true,
            "schema": {
              "originalRef": "QuestionQueryRequest",
              "$ref": "#/definitions/QuestionQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«QuestionVO»»",
              "$ref": "#/definitions/BaseResponse«Page«QuestionVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«QuestionVO»»",
              "$ref": "#/definitions/BaseResponse«Page«QuestionVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/question/question_submit/do": {
      "post": {
        "tags": [
          "question-controller"
        ],
        "summary": "doQuestionSubmit",
        "operationId": "doQuestionSubmitUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "questionSubmitAddRequest",
            "description": "questionSubmitAddRequest",
            "required": true,
            "schema": {
              "originalRef": "QuestionSubmitAddRequest",
              "$ref": "#/definitions/QuestionSubmitAddRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«long»",
              "$ref": "#/definitions/BaseResponse«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«long»",
              "$ref": "#/definitions/BaseResponse«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/question/question_submit/list/page": {
      "post": {
        "tags": [
          "question-controller"
        ],
        "summary": "listQuestionSubmitByPage",
        "operationId": "listQuestionSubmitByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "questionSubmitQueryRequest",
            "description": "questionSubmitQueryRequest",
            "required": true,
            "schema": {
              "originalRef": "QuestionSubmitQueryRequest",
              "$ref": "#/definitions/QuestionSubmitQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«QuestionSubmitVO»»",
              "$ref": "#/definitions/BaseResponse«Page«QuestionSubmitVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«QuestionSubmitVO»»",
              "$ref": "#/definitions/BaseResponse«Page«QuestionSubmitVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/question/update": {
      "post": {
        "tags": [
          "question-controller"
        ],
        "summary": "updateQuestion",
        "operationId": "updateQuestionUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "questionUpdateRequest",
            "description": "questionUpdateRequest",
            "required": true,
            "schema": {
              "originalRef": "QuestionUpdateRequest",
              "$ref": "#/definitions/QuestionUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/api/user/get/login": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getLoginUser",
        "operationId": "getLoginUserUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«LoginUserVO»",
              "$ref": "#/definitions/BaseResponse«LoginUserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«LoginUserVO»",
              "$ref": "#/definitions/BaseResponse«LoginUserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
  },
  "definitions": {
...
    "BaseResponse«Question»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "originalRef": "Question",
          "$ref": "#/definitions/Question"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«Question»"
    },
...
    "BaseResponse«boolean»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«boolean»"
    },
...
    "BaseResponse«long»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«long»"
    },
...
    "DeleteRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "DeleteRequest"
    },
...
    "QuestionAddRequest": {
      "type": "object",
      "properties": {
        "answer": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "judgeCase": {
          "type": "array",
          "items": {
            "originalRef": "JudgeCase",
            "$ref": "#/definitions/JudgeCase"
          }
        },
        "judgeConfig": {
          "originalRef": "JudgeConfig",
          "$ref": "#/definitions/JudgeConfig"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        }
      },
      "title": "QuestionAddRequest"
    },
...
  },
...
}